requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
google-generativeai==0.3.2
openai==1.6.1
python-dateutil==2.8.2

